refactor: refactor and improve example/combine-segs with good test coverage
